+++ grizzly/tasks/print.py
@@ -9,7 +9,7 @@ from typing import TYPE_CHECKING, Any, Callable
 
 from jinja2 import Template
 
-from ..types import GrizzlyTask
+from . import GrizzlyTask
 
 if TYPE_CHECKING:  # pragma: no cover
     from ..scenarios import GrizzlyScenario
