+++ grizzly/types.py
@@ -1,15 +1,10 @@
-from abc import ABCMeta
 from enum import Enum
-from typing import TYPE_CHECKING, Callable, Optional, Tuple, Any, Union, Dict, TypeVar, List, Type, Generic, Set, cast
+from typing import Callable, Optional, Tuple, Any, Union, Dict, TypeVar, List, Type, Generic, Set, cast
 from importlib import import_module
 
 from aenum import Enum as AdvancedEnum, NoAlias
 from locust.clients import ResponseContextManager
 from gevent.lock import Semaphore
-
-if TYPE_CHECKING:  # pragma: no cover
-    from .scenarios import GrizzlyScenario
-    from .context import GrizzlyContextScenario
 
 
 class ResponseTarget(Enum):
@@ -251,13 +246,3 @@ class GrizzlyDict(dict):
             value = caster(value)
 
         super().__setitem__(key, value)
-
-
-class GrizzlyTask(metaclass=ABCMeta):
-    scenario: 'GrizzlyContextScenario'
-
-    def __call__(self) -> Callable[['GrizzlyScenario'], Any]:
-        raise NotImplementedError(f'{self.__class__.__name__} has not implemented "implementation"')
-
-    def get_templates(self) -> List[str]:
-        return []
